pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.11.0",
    "discord-py>=2.6.3",
    "flask>=3.0.0",
    "google-genai>=1.41.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sift-stack-py>=0.9.1",
]
